+++ include/yaml_blueprints.php
@@ -1,7 +1,6 @@
 <?php
-
-//YAML - blueprint related functions
-include_once('spyc/Spyc.php');
+//YAML - graphics related functions
+//include_once('spyc/Spyc.php');
 include_once('yaml_common.php');
 
 function updateYamlBlueprints($silent=true) {
@@ -55,7 +54,12 @@ function updateYamlBlueprints($silent=true) {
     db_uquery($createyamlBlueprintSkills);
     
     if (!$silent) echo('loading YAML...');
-    $blueprints = Spyc::YAMLLoad($file);
+    //switching from Spyc to YAML PECL module
+    $blueprints = yaml_parse_wrapper($file);
+    if ($blueprints===FALSE) die("yaml_parse_file failed for $file");
+    
+    //var_dump($blueprints);
+    
     if (!empty($blueprints)) {
         db_uquery("TRUNCATE TABLE `$LM_EVEDB`.`yamlBlueprintTypes`;");
         db_uquery("TRUNCATE TABLE `$LM_EVEDB`.`yamlBlueprintProducts`;");
@@ -75,13 +79,14 @@ function updateYamlBlueprints($silent=true) {
         //var_dump($blueprint);
         //second, walk activities
         foreach ($blueprint['activities'] as $activityName => $activity) {
+            //var_dump($activity);
             $activityID=yaml_activity2ID($activityName);
             $time=yaml_prepare($activity['time'],0);
             
             //walk new (post-Phoebe) flat YAML output
-            foreach($activity as $key => $entry) {
-                if (!is_numeric($key)) $currentList=$key; else {
-                    switch($currentList) {
+            foreach($activity as $key => $entries) {
+                if (is_array($entries)) foreach($entries as $entry) {
+                    switch($key) {
                         case 'materials':
                             $materialTypeID=yaml_prepare($entry['typeID']);
                             $quantity=yaml_prepare($entry['quantity'],0);
@@ -153,7 +158,7 @@ function recreateLegacyTables() {
     NULL AS `parentBlueprintTypeID`,
     manu.`productTypeID`,
     manu.`time` AS `productionTime`,
-    imt.`metaGroupID` AS `techLevel`,
+    COALESCE(dgm.`valueInt`,dgm.`valueFloat`) AS `techLevel`,
     te.`time` AS `researchProductivityTime`,
     me.`time` AS `researchMaterialTime`,
     copy.`time` AS `researchCopyTime`,
@@ -175,6 +180,8 @@ function recreateLegacyTables() {
       ON ybt.`blueprintTypeID`=inv.`blueprintTypeID` AND inv.`activityID`=8
     LEFT JOIN `$LM_EVEDB`.`invMetaTypes` imt
       ON manu.`productTypeID`=imt.`typeID`
+    LEFT JOIN `$LM_EVEDB`.`dgmTypeAttributes` dgm
+      ON manu.`productTypeID`=dgm.`typeID` AND dgm.`attributeID`=422
     WHERE TRUE;";
     
     db_uquery($insertinvBlueprintTypes);
@@ -271,7 +278,10 @@ function updateYamlBlueprints_pre_Phoebe($silent=true) {
     db_uquery($createyamlBlueprintSkills);
     
     if (!$silent) echo('loading YAML...');
-    $blueprints = Spyc::YAMLLoad($file);
+    //switching from Spyc to YAML PECL module
+    $blueprints = yaml_parse_wrapper($file);
+    if ($blueprints===FALSE) die("yaml_parse_file failed for $file");
+    
     if (!empty($blueprints)) {
         db_uquery("TRUNCATE TABLE `$LM_EVEDB`.`yamlBlueprintTypes`;");
         db_uquery("TRUNCATE TABLE `$LM_EVEDB`.`yamlBlueprintProducts`;");
